consolidate has_wcs logic
