+++ jdaviz/app.py
@@ -14,7 +14,6 @@ from astropy.nddata import CCDData, NDData
 from astropy.io import fits
 from astropy.time import Time
 from astropy.utils.decorators import deprecated
-from astropy.wcs.wcsapi import BaseHighLevelWCS
 from echo import CallbackProperty, DictCallbackProperty, ListCallbackProperty
 from ipygoldenlayout import GoldenLayout
 from ipysplitpanes import SplitPanes
@@ -58,7 +57,7 @@ from jdaviz.core.style_widget import StyleWidget
 from jdaviz.core.registries import (tool_registry, tray_registry, viewer_registry,
                                     data_parser_registry)
 from jdaviz.core.tools import ICON_DIR
-from jdaviz.utils import SnackbarQueue, alpha_index, MultiMaskSubsetState
+from jdaviz.utils import SnackbarQueue, alpha_index, data_has_valid_wcs, MultiMaskSubsetState
 from ipypopout import PopoutButton
 
 __all__ = ['Application', 'ALL_JDAVIZ_CONFIGS']
@@ -554,11 +553,8 @@ class Application(VuetifyTemplate, HubListener):
             )
 
         # only re-center the viewer if all data layers have WCS:
-        data_has_wcs = [
-            hasattr(d, 'coords') and isinstance(d.coords, BaseHighLevelWCS)
-            for d in viewer.data()
-        ]
-        if all(data_has_wcs):
+        has_wcs_per_data = [data_has_valid_wcs(d) for d in viewer.data()]
+        if all(has_wcs_per_data):
             # re-center the viewer on previous location.
             viewer.center_on(sky_cen)
 
@@ -2391,7 +2387,7 @@ class Application(VuetifyTemplate, HubListener):
             'locked': False,
             'ndims': data.ndim,
             'type': typ,
-            'has_wcs': wcsaxes is not None,
+            'has_wcs': data_has_valid_wcs(data),
             'meta': {k: v for k, v in data.meta.items() if _expose_meta(k)},
             'children': []}
